src/app/prediction_service/requirements.txt
# local package
#-e .

# external requirements
click
Sphinx
coverage
#awscli>=1.29.0
flake8
pathlib 
joblib

# pour le modele
pandas
numpy
imblearn
scikit-learn==1.3.2
imbalanced-learn

# pour l api
uvicorn
fastapi 
pydantic 
#logging
passlib
evidently

#pour test
pytest
httpx

#pour la base de données
psycopg2-binary